dynamodb & refactoring

+++ src/Service/Telegram/Bot/Conversation/TelegramBotConversationManager.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace App\Service\Telegram\Bot\Conversation;
 
 use App\Entity\Telegram\TelegramBotConversation;
-use App\Entity\Telegram\TelegramBotConversationState;
-use App\Entity\Telegram\TelegramBotStoppedConversation;
+use App\Model\Telegram\TelegramBotConversationState;
 use App\Repository\Telegram\Bot\TelegramBotConversationRepository;
 use App\Service\ORM\EntityManager;
 use App\Service\Telegram\Bot\Group\TelegramBotGroupRegistry;
@@ -14,6 +13,8 @@ use App\Service\Telegram\Bot\TelegramBot;
 use App\Service\Telegram\Bot\TelegramBotAwareHelper;
 use App\Service\Telegram\Bot\TelegramBotChatProvider;
 use App\Service\Util\Array\ArrayNullFilter;
+use DateTimeImmutable;
+use Psr\Log\LoggerInterface;
 use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
@@ -31,6 +32,8 @@ class TelegramBotConversationManager
         private readonly ArrayNullFilter $arrayNullFilter,
         private readonly TelegramBotGroupRegistry $telegramBotGroupRegistry,
         private readonly TelegramBotChatProvider $telegramBotChatProvider,
+        private readonly LoggerInterface $logger,
+        private readonly bool $saveRequests = false,
     )
     {
     }
@@ -40,13 +43,24 @@ class TelegramBotConversationManager
         $messengerUser = $bot->getMessengerUser();
         $chatId = $this->telegramBotChatProvider->getTelegramChatByUpdate($bot->getUpdate())?->getId();
 
+//        $this->logger?->debug(__METHOD__, [
+//            'messengerUser' => $messengerUser,
+//            'chatId' => $chatId,
+//        ]);
+
         if ($messengerUser === null || $chatId === null) {
             return null;
         }
 
         $hash = $this->createTelegramConversationHash($messengerUser->getId(), $chatId, $bot->getEntity()->getId());
+        $conversation = $this->telegramBotConversationRepository->findOneNonDeletedByHash($hash);
 
-        return $this->telegramBotConversationRepository->findOneByHash($hash);
+        $this->logger?->debug(__METHOD__, [
+            'hash' => $hash,
+            'conversation' => $conversation,
+        ]);
+
+        return $conversation;
     }
 
     public function startTelegramConversation(TelegramBot $bot, string $class): void
@@ -58,7 +72,7 @@ class TelegramBotConversationManager
 
     public function createTelegramConversationHash(string $messengerUserId, int $chatId, string $botId): string
     {
-        return $messengerUserId . '-' . $chatId . '-' . $botId;
+        return md5($messengerUserId . '-' . $chatId . '-' . $botId);
     }
 
     public function createTelegramConversation(
@@ -97,9 +111,9 @@ class TelegramBotConversationManager
         $this->executeConversation($bot, $entity, $method);
     }
 
-    public function continueTelegramConversation(TelegramBot $bot, TelegramBotConversation $entity): void
+    public function continueTelegramConversation(TelegramBot $bot, TelegramBotConversation $conversation): void
     {
-        $this->executeConversation($bot, $entity, 'invoke');
+        $this->executeConversation($bot, $conversation, 'invoke');
     }
 
     public function denormalizeState(?array $state, string $class): TelegramBotConversationState
@@ -131,17 +145,12 @@ class TelegramBotConversationManager
 
     public function stopTelegramConversation(TelegramBotConversation $entity): void
     {
-        $stopped = new TelegramBotStoppedConversation(
-            $entity->getMessengerUserId(),
-            $entity->getChatId(),
-            $entity->getBotId(),
-            $entity->getClass(),
-            $entity->getState(),
-            $entity->getCreatedAt()
-        );
-        $this->entityManager->persist($stopped);
-
-        $this->entityManager->remove($entity);
+        if ($this->saveRequests) {
+            $entity->setDeletedAt(new DateTimeImmutable());
+            $entity->setExpireAt((new DateTimeImmutable())->setTimestamp(time() + 7 * 24 * 60 * 60));
+        } else {
+            $this->entityManager->remove($entity);
+        }
     }
 
     public function executeConversation(
@@ -164,6 +173,13 @@ class TelegramBotConversationManager
         $state = $this->normalizeState($conversation->getState());
 
         $entity->setState($state);
+
+//        $this->logger->critical('-------------------->' . __METHOD__, [
+//            'conv_hash' => $entity->getHash(),
+//            'conv_deleted_at' => $entity->getDeletedAt(),
+//            'conv_expire_at' => $entity->getExpireAt(),
+//            'state' => $state,
+//        ]);
 
         return $conversation;
     }
